refactor: DTO 폴더 구조 개선 (request/response 폴더로 분리)
