+++ src/main/java/com/example/eat_together/domain/cart/dto/request/CartItemRequestDto.java
@@ -1,11 +1,13 @@
-package com.example.eat_together.domain.cart.dto;
+package com.example.eat_together.domain.cart.dto.request;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
+@Setter
 public class CartItemRequestDto {
 
     @NotNull(message = "메뉴 ID는 필수입니다.")
